pyproject.toml
[project]
name = "beautycity"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.13.*"
dependencies = [
    "dj-database-url>=3.1.0",
    "django==5.2.10",
    "django-environ>=0.12.0",
    "django-phonenumber-field[phonenumberslite]>=8.4.0",
    "djangorestframework>=3.16.1",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "ruff>=0.14.13",
]
